Fix: extract browser integration fake path
